Time: 3 ms (37.64%), Space: 113.8 MB (37.06%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.cpp
@@ -4,14 +4,14 @@ public:
         
         // find contigious subarray sums
         int i = 0;
-        double sum = 0;
+        int sum = 0;
         while(i<k){
             sum += nums[i];
             i++;
         }
         int j = 0;
         int n = nums.size();
-        double max = sum;
+        int max = sum;
         while(i < n){
             sum += nums[i];
             sum -= nums[j];
